refactor: componentizando marcadores do mapa
